requirements.txt
discord.py
redis
imgurpython
